server/requirements/base.txt
# --- Core Framework ---
fastapi==0.115.14
uvicorn[standard]==0.32.0

# --- Database ---
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# --- Auth & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.2
python-dotenv==1.0.1

# --- Realtime & Chat ---
websockets==13.1
redis==5.1.1
aiofiles==24.1.0

# --- Email/Image/OCR ---
httpx==0.28.1
aiohttp==3.10.11
Pillow==11.0.0
# tesserocr==2.8.0

# --- Geolocation ---
geopy==2.4.1
ip2geotools==0.1.6

# --- Environment ---
pydantic 
python-dotenv

python-jose